FungusScript can now be made into a prefab (see notes!)
All sequence and command components are now added to the main
FungusScript object. This change enables you to make a prefab out of
any FungusScript. FungusScript prefabs can now be edited using the
Fungus Script window. You still need to instantiate a Fungus Script
prefab in order to run it.