Removed delay property from Invoke Method
Invoke Method supports calling coroutine functions with the option to
stop or wait until the coroutine exits before continuing with the next
command. This conflicts with the concept of delaying invocation of the
method until later and continuing immediately. Rather than have a hard
to understand interaction it’s better to just let users add a Wait
command before Invoke Command when they want to introduce a delay.