Add files via upload
So I've managed to write a script that allows you to write Fungus variable values into Playmaker global variables, but here's the catch. The script doesn't read in variables from a flowchart to copy over, it just reads them in from the inspector. I've managed to find a way to read in Strings from the flowchart, but not ints, floats or booleans. I checked the tutorial video[1] for reading variable values from the flowchart, but the tutorial only deals with hard coding the name of the variable you want to read in. Any advice that could help me fix this would be greatly appreciated.

[1]https://youtu.be/uRh7INbed2I?t=1219