New commands for additive Scenes in Unity.
- SceneActivateAdditive
Activates a loaded scene. Just needs the scene name - the Scene MUST be added in Build Settings.

- SceneLoadAdditive
Loads a scene. Optionally activates it, and calls a Fungus block after activation.

- SceneRemoveAddtitive
Remove the added scene. This has to be called, when a Scene is not needed anymore.

- SceneEditor
Helper to see the blocks in a scene, when using SceneLoadAdditive