AutoStart and call cleanup
Simplify Call logic to track where onComplete delegate are set
Add AutoStart EventHandler to replace use of IsStartingPoint logic in ProgressMarker and SaveManager, as a command that causes a block to auto run is incongruent.
Add addition fixed data per block in BlockData saving, prev and next command index.
Add visitor for encode and decode of block command and event handler, to be used for more generic storage of data, unused at this commit.