addressing issue #929
This issue seemed simple to tackle and the feature would be nice to add depending on what audio you use. The dopplering effect is minor with the built-in sound clips, but amplified with things like typewriters.

* Added a Lerp Writer Audio bool.
  * When enabled (default), the WriterAudio uses the default behaviour of using MoveTowards for volume.
  * When disabled, the WriterAudio sets the audio volume to the target volume while writing, removing the lerp effect.

This setting is enabled by default on the say dialogue prefab Fungus spawns by default and therefore cannot be set at runtime.