Select command upon adding
Hopefully fixes #994. This needs to be tested upon returning home. (This was written at work)

Essentially, once the command is added and the other commands are cleared, the new command is automatically selected, allowing editing right away. It also means that adding commands in succession means additional commands do not go to the bottom of the block, so long as the first command was not deselected.