Added flowchart locking
Added functionality so a selected flowchart can be locked. This prevents the commands from being edited and reordered, block data editing, and block movement. Navigating and selecting blocks and commands is still allowed.

This is also on a per-flowchart basis, so locked prefabs can be made and those prefab instances unlocked. Vice versa also works, with unlocked flowcharts in a scene and locked prefabs.

This does work with legacy menus as well.