Laying out groundwork for restructure in how save data is defined
I feel that rather than relying on StringPairs to be the save data units, there should be different classes handling the job for different types of state.

I find this approach a lot more extensible, flexible, and as a result, more user-friendly.

Of course, this means that the unit tests may need to be restructured. New tests will be written as the restructure continues.