Refactored all InfoStructs as ISaveUnits
I realized that given the purpose of save units, there is no need to have info structs and save units be separate classes. The latter were just working as needless wrappers for the former.

Some refactoring of the stage-saving tests was needed since the portrait state containers are now structs instead of classes.